Breakout grpc types into own file
